testing

+++ lib/LightSensor/src/LightSensor.cpp
@@ -4,9 +4,9 @@ LightSensor::LightSensor(uint8_t pin)
 {
     this->pin = pin;
     pinMode(this->pin, INPUT);
+    //pinMode(A1,INPUT);
 }
-unsigned short LightSensor::getLightLevel()
+int LightSensor::getLightLevel()
 {
-    return analogRead(this->pin);
-
+    return analogRead(this->pin /*A1*/);
 }
